+++ src/plugins/net_plugin/CMakeLists.txt
@@ -59,6 +59,7 @@ target_link_libraries(net_plugin
 target_include_directories(net_plugin PUBLIC
         "${CMAKE_CURRENT_SOURCE_DIR}/../../../lib/appbase"
         "${CMAKE_CURRENT_SOURCE_DIR}/../../../lib/log"
+        "${CMAKE_CURRENT_SOURCE_DIR}/../../../lib/gruut-utils/src"
         "${GRUUT_INSTALL_DIR}/include"
         ${Boost_INCLUDE_DIR}
         )
