+++ src/plugins/net_plugin/include/msg_handler.hpp
@@ -3,7 +3,7 @@
 #include "../../../../lib/json/include/json.hpp"
 #include "../config/include/message.hpp"
 #include "../config/include/network_config.hpp"
-#include "../../../gruut-utils/src/lz4_compressor.hpp"
+#include "../../../../lib/gruut-utils/src/lz4_compressor.hpp"
 #include "../../channel_interface/include/channel_interface.hpp"
 
 using namespace grpc;
